requirements.txt
python-telegram-bot
feedgen
boto3